kauthapp/requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2; python_version >= '3.7'
certifi==2022.9.14; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
django==4.1.1
django-registration-redux==2.11
djangorestframework==3.13.1
drf-standardized-errors==0.12.1
gunicorn==20.1.0
idna==3.4; python_version >= '3.5'
oauthlib==3.2.1; python_version >= '3.6'
psycopg2-binary==2.9.3
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
setuptools==65.3.0; python_version >= '3.7'
sqlparse==0.4.2; python_version >= '3.5'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
whitenoise==6.2.0
